+++ coaster/logger.py
@@ -10,12 +10,15 @@ this is done automatically for you.
 from datetime import datetime, timedelta
 from io import StringIO
 from pprint import pprint
-from typing import Dict
+from threading import Lock
+from typing import Any, Dict
 import logging.handlers
 import re
+import textwrap
 import traceback
 
 from flask import escape, g, request, session
+from flask.config import Config
 
 import requests
 
@@ -41,20 +44,23 @@ _filter_re = re.compile(
     |email
     |phone)
     ''',
-    re.I | re.X,
+    re.IGNORECASE | re.VERBOSE,
 )
 
 # global var as lazy in-memory cache
-error_throttle_timestamp_sms: Dict[str, datetime] = {}
 error_throttle_timestamp_slack: Dict[str, datetime] = {}
 error_throttle_timestamp_telegram: Dict[str, datetime] = {}
 
 
 class FilteredValueIndicator:
+    """Represent a filtered value."""
+
     def __str__(self):
+        """Filter str."""
         return '[Filtered]'
 
     def __repr__(self):
+        """Filter repr."""
         return '[Filtered]'
 
 
@@ -62,7 +68,22 @@ class FilteredValueIndicator:
 filtered_value_indicator = FilteredValueIndicator()
 
 
+class RepeatValueIndicator:
+    """Represent a repeating value."""
+
+    def __init__(self, key):
+        """Init with key."""
+        self.key = key
+
+    def __repr__(self):
+        """Return representation."""
+        return '<same as prior "%s">' % self.key
+
+    __str__ = __repr__
+
+
 def filtered_value(key, value):
+    """Find and mask sensitive values based on key names."""
     if isinstance(key, str) and _filter_re.search(key):
         return filtered_value_indicator
     if isinstance(value, str):
@@ -76,22 +97,23 @@ def pprint_with_indent(dictlike, outfile, indent=4):
     pprint(  # NOQA: T003
         {key: filtered_value(key, value) for key, value in dictlike.items()}, out
     )
-    # textwrap.indent would have been simpler but is not present in Python 2.7
-    outfile.write(
-        '\n'.join((' ' * indent) + line for line in out.getvalue().split('\n'))
-    )
+    outfile.write(textwrap.indent(out.getvalue(), ' ' * indent))
     out.close()
 
 
 class LocalVarFormatter(logging.Formatter):
-    """
-    Custom log formatter that logs the contents of local variables in the stack frame.
-    """
+    """Log the contents of local variables in the stack frame."""
+
+    def __init__(self, *args, **kwargs):
+        """Init formatter."""
+        super().__init__(*args, **kwargs)
+        self.lock = Lock()
 
     def format(self, record):  # NOQA: A003
         """
-        Format the specified record as text. Overrides
-        :meth:`logging.Formatter.format` to remove cache of
+        Format the specified record as text.
+
+        Overrides :meth:`logging.Formatter.format` to remove cache of
         :attr:`record.exc_text` unless it was produced by this formatter.
         """
         if (
@@ -102,7 +124,8 @@ class LocalVarFormatter(logging.Formatter):
             record.exc_text = None
         return super().format(record)
 
-    def formatException(self, ei):  # NOQA: N802
+    def formatException(self, ei) -> str:  # NOQA: N802
+        """Render a stack trace with local variables in each stack frame."""
         tb = ei[2]
         while True:
             if not tb.tb_next:
@@ -117,36 +140,56 @@ class LocalVarFormatter(logging.Formatter):
         sio = StringIO()
         traceback.print_exception(ei[0], ei[1], ei[2], None, sio)
 
-        print('\n----------\n', file=sio)  # NOQA: T001
-        # XXX: The following text is used as a signature in :meth:`format` above
-        print("Stack frames (most recent call first):", file=sio)  # NOQA: T001
-        for frame in stack:
-            print('\n----\n', file=sio)  # NOQA: T001
-            print(  # NOQA: T001
-                "Frame %s in %s at line %s"
-                % (frame.f_code.co_name, frame.f_code.co_filename, frame.f_lineno),
-                file=sio,
-            )
-            for key, value in list(frame.f_locals.items()):
-                print("\t%20s = " % key, end=' ', file=sio)  # NOQA: T001
-                try:
-                    print(repr(filtered_value(key, value)), file=sio)  # NOQA: T001
-                except:  # NOQA
-                    # We need a bare except clause because this is the exception
-                    # handler. It can't have exceptions of its own.
-                    print("<ERROR WHILE PRINTING VALUE>", file=sio)  # NOQA: T001
+        with self.lock:
+            # Monkey-patch Flask Config's __repr__ to not dump sensitive config. This
+            # can happen when a Jinja2 template is part of the stack, as templates get
+            # app config. We perform this patch and dump within a lock to ensure no
+            # conflict with a parallel stack dump -- which could otherwise restore the
+            # original __repr__ while this is still dumping.
+            original_config_repr = Config.__repr__
+            Config.__repr__ = lambda self: '<Config [FILTERED]>'  # type: ignore[assignment]
+            value_cache: Dict[Any, str] = {}
+
+            print('\n----------\n', file=sio)  # NOQA: T001
+            # XXX: The following text is used as a signature in :meth:`format` above
+            print("Stack frames (most recent call first):", file=sio)  # NOQA: T001
+            for frame in stack:
+                print('\n----\n', file=sio)  # NOQA: T001
+                print(  # NOQA: T001
+                    "Frame %s in %s at line %s"
+                    % (frame.f_code.co_name, frame.f_code.co_filename, frame.f_lineno),
+                    file=sio,
+                )
+                for key, value in list(frame.f_locals.items()):
+                    idvalue = id(value)
+                    if idvalue in value_cache:
+                        value = RepeatValueIndicator(value_cache[idvalue])
+                    else:
+                        value_cache[idvalue] = key
+                    print("\t%20s = " % key, end=' ', file=sio)  # NOQA: T001
+                    try:
+                        print(repr(filtered_value(key, value)), file=sio)  # NOQA: T001
+                    except:  # NOQA
+                        # We need a bare except clause because this is the exception
+                        # handler. It can't have exceptions of its own.
+                        print("<ERROR WHILE PRINTING VALUE>", file=sio)  # NOQA: T001
+
+            del value_cache
+            Config.__repr__ = original_config_repr  # type: ignore[assignment]
 
         if request:
             print('\n----------\n', file=sio)  # NOQA: T001
             print("Request context:", file=sio)  # NOQA: T001
             request_data = {
-                'form': request.form,
-                'args': request.args,
-                'cookies': request.cookies,
-                'stream': request.stream,
+                'form': {
+                    k: filtered_value(k, v)
+                    for k, v in request.form.to_dict(flat=False).items()
+                },
+                'args': {
+                    k: filtered_value(k, v)
+                    for k, v in request.args.to_dict(flat=False).items()
+                },
                 'headers': request.headers,
-                'data': request.data,
-                'files': request.files,
                 'environ': request.environ,
                 'method': request.method,
                 'blueprint': request.blueprint,
@@ -189,87 +232,17 @@ class LocalVarFormatter(logging.Formatter):
         return s
 
 
-class SMSHandler(logging.Handler):
-    """
-    Custom logging handler to send SMSes to admins
-    """
-
-    def __init__(
-        self,
-        app_name,
-        exotel_sid,
-        exotel_token,
-        exotel_from,
-        twilio_sid,
-        twilio_token,
-        twilio_from,
-        phonenumbers,
-    ):
-        logging.Handler.__init__(self)
-        self.app_name = app_name
-        self.phonenumbers = phonenumbers
-        self.exotel_sid = exotel_sid
-        self.exotel_token = exotel_token
-        self.exotel_from = exotel_from
-        self.twilio_sid = twilio_sid
-        self.twilio_token = twilio_token
-        self.twilio_from = twilio_from
-
-    def emit(self, record):
-        throttle_key = (record.module, record.lineno)
-        if throttle_key not in error_throttle_timestamp_sms or (
-            (datetime.utcnow() - error_throttle_timestamp_sms[throttle_key])
-            > timedelta(minutes=5)
-        ):
-            msg = "{message}: {info}".format(
-                message=record.message,
-                info=repr(record.exc_info[1]) if record.exc_info else '',
-            )
-            for phonenumber in self.phonenumbers:
-                self.sendsms(
-                    phonenumber,
-                    "Error in {name}. {msg}. "
-                    "Please check your email for details".format(
-                        name=self.app_name, msg=msg
-                    ),
-                )
-            error_throttle_timestamp_sms[throttle_key] = datetime.utcnow()
-
-    def sendsms(self, number, message):
-        try:
-            if number.startswith('+91'):  # Indian
-                requests.post(
-                    'https://twilix.exotel.in/v1/Accounts/{sid}/Sms/send.json'.format(
-                        sid=self.exotel_sid
-                    ),
-                    auth=(self.exotel_sid, self.exotel_token),
-                    data={'From': self.exotel_from, 'To': number, 'Body': message},
-                )
-            else:
-                requests.post(
-                    'https://api.twilio.com/2010-04-01/Accounts/{sid}/Messages.json'.format(
-                        sid=self.twilio_sid
-                    ),
-                    auth=(self.twilio_sid, self.twilio_token),
-                    data={'From': self.twilio_from, 'To': number, 'Body': message},
-                )
-        except:  # NOQA  # nosec
-            # We need a bare except clause because this is the exception handler.
-            # It can't have exceptions of its own.
-            pass
-
-
 class SlackHandler(logging.Handler):
-    """
-    Custom logging handler to post error reports to Slack.
-    """
+    """Custom logging handler to post error reports to Slack."""
 
     def __init__(self, app_name, webhooks):
+        """Init handler."""
         super().__init__()
         self.app_name = app_name
         self.webhooks = webhooks
 
     def emit(self, record):
+        """Emit an event."""
         throttle_key = (record.module, record.lineno)
         if throttle_key not in error_throttle_timestamp_slack or (
             (datetime.utcnow() - error_throttle_timestamp_slack[throttle_key])
@@ -337,17 +310,17 @@ class SlackHandler(logging.Handler):
 
 
 class TelegramHandler(logging.Handler):
-    """
-    Custom logging handler to report errors to a Telegram chat
-    """
+    """Custom logging handler to report errors to a Telegram chat."""
 
     def __init__(self, app_name, chatid, apikey):
+        """Init handler."""
         super().__init__()
         self.app_name = app_name
         self.chatid = chatid
         self.apikey = apikey
 
     def emit(self, record):
+        """Emit an event."""
         throttle_key = (record.module, record.lineno)
         if throttle_key not in error_throttle_timestamp_telegram or (
             (datetime.utcnow() - error_throttle_timestamp_telegram[throttle_key])
@@ -382,9 +355,10 @@ class TelegramHandler(logging.Handler):
 
 def init_app(app):
     """
-    Enables logging for an app using :class:`LocalVarFormatter`. Requires the
-    app to be configured and checks for the following configuration parameters.
-    All are optional:
+    Enable logging for an app using :class:`LocalVarFormatter`.
+
+    Requires the app to be configured and checks for the following configuration
+    parameters. All are optional:
 
     * ``LOGFILE``: Name of the file to log to (default ``error.log``)
     * ``LOGFILE_LEVEL``: Logging level to use for file logger (default `WARNING`)
@@ -395,14 +369,8 @@ def init_app(app):
     * ``MAIL_USERNAME`` and ``MAIL_PASSWORD``: SMTP credentials, if required
     * ``SLACK_LOGGING_WEBHOOKS``: If present, will send error logs to all specified
         Slack webhooks
-    * ``ADMIN_NUMBERS``: List of mobile numbers of admin to send SMS alerts. Requires
-        the following values too
-    * ``SMS_EXOTEL_SID``: Exotel SID for Indian numbers (+91 prefix)
-    * ``SMS_EXOTEL_TOKEN``: Exotel token
-    * ``SMS_EXOTEL_FROM``: Exotel sender's number
-    * ``SMS_TWILIO_SID``: Twilio SID for non-Indian numbers
-    * ``SMS_TWILIO_TOKEN``: Twilio token
-    * ``SMS_TWILIO_FROM``: Twilio sender's number
+    * ``TELEGRAM_ERROR_CHATID`` and ``TELEGRAM_ERROR_APIKEY``: If present, will use the
+        specified API key to post a message to the specified chat
 
     Format for ``SLACK_LOGGING_WEBHOOKS``::
 
@@ -412,10 +380,6 @@ def init_app(app):
             }]
 
     """
-    if not app.debug:
-        # Downgrade from default WARNING level to INFO
-        app.logger.setLevel(logging.INFO)
-
     formatter = LocalVarFormatter(
         '%(asctime)s - %(module)s.%(funcName)s:%(lineno)s - %(levelname)s - %(message)s'
     )
@@ -430,35 +394,6 @@ def init_app(app):
         file_handler.setFormatter(formatter)
         file_handler.setLevel(app.config.get('LOGFILE_LEVEL', logging.WARNING))
         app.logger.addHandler(file_handler)
-
-    if app.config.get('ADMIN_NUMBERS') and all(
-        key in app.config
-        for key in [
-            'SMS_EXOTEL_SID',
-            'SMS_EXOTEL_TOKEN',
-            'SMS_EXOTEL_FROM',
-            'SMS_TWILIO_SID',
-            'SMS_TWILIO_TOKEN',
-            'SMS_TWILIO_FROM',
-        ]
-    ):
-
-        # A little trickery because directly creating
-        # an SMSHandler object didn't work
-        logging.handlers.SMSHandler = SMSHandler
-
-        sms_handler = logging.handlers.SMSHandler(
-            app_name=app.config.get('SITE_ID') or app.name,
-            exotel_sid=app.config['SMS_EXOTEL_SID'],
-            exotel_token=app.config['SMS_EXOTEL_TOKEN'],
-            exotel_from=app.config['SMS_EXOTEL_FROM'],
-            twilio_sid=app.config['SMS_TWILIO_SID'],
-            twilio_token=app.config['SMS_TWILIO_TOKEN'],
-            twilio_from=app.config['SMS_TWILIO_FROM'],
-            phonenumbers=app.config['ADMIN_NUMBERS'],
-        )
-        sms_handler.setLevel(logging.ERROR)
-        app.logger.addHandler(sms_handler)
 
     if app.config.get('SLACK_LOGGING_WEBHOOKS'):
         logging.handlers.SlackHandler = SlackHandler
@@ -483,11 +418,7 @@ def init_app(app):
         app.logger.addHandler(telegram_handler)
 
     if app.config.get('ADMINS'):
-        # MAIL_DEFAULT_SENDER is the new setting for default mail sender in Flask-Mail
-        # DEFAULT_MAIL_SENDER is the old setting. We look for both
-        mail_sender = app.config.get('MAIL_DEFAULT_SENDER') or app.config.get(
-            'DEFAULT_MAIL_SENDER', 'logs@example.com'
-        )
+        mail_sender = app.config.get('MAIL_DEFAULT_SENDER', 'logs@example.com')
         if isinstance(mail_sender, (list, tuple)):
             mail_sender = mail_sender[1]  # Get email from (name, email)
         if app.config.get('MAIL_USERNAME') and app.config.get('MAIL_PASSWORD'):